[added] tumbletokneeling, tumbletoprone, tumbletostanding, aim commands [added] bank terminals to player bank structures
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2294 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/creature/commands/CombatQueueCommand.h
@@ -36,7 +36,7 @@ protected:
 	int nextAttackDelayChance;
 	int durationStateTime;
 
-	float dotDuration;
+	uint32 dotDuration;
 	uint64 dotType;
 	uint8 dotPool;
 	uint32 dotStrength;
